Fix issue with 255+ rows used in octamed

+++ synth/mod_player.h
@@ -30,9 +30,9 @@ typedef struct ModPlayer ModPlayer;
  * Called when song order, pattern number, or row changes
  * order: Current position in song order (0-127)
  * pattern: Actual pattern number being played
- * row: Current row within pattern (0-63)
+ * row: Current row within pattern (0-65535)
  */
-typedef void (*ModPlayerPositionCallback)(uint8_t order, uint8_t pattern, uint8_t row, void* user_data);
+typedef void (*ModPlayerPositionCallback)(uint8_t order, uint8_t pattern, uint16_t row, void* user_data);
 
 /**
  * MOD sample structure
@@ -142,7 +142,7 @@ void mod_player_set_loop_range(ModPlayer* player, uint8_t start_pattern, uint8_t
 /**
  * Get current position
  */
-void mod_player_get_position(const ModPlayer* player, uint8_t* pattern, uint8_t* row);
+void mod_player_get_position(const ModPlayer* player, uint8_t* pattern, uint16_t* row);
 
 /**
  * Set position change callback
@@ -154,7 +154,7 @@ void mod_player_set_position_callback(ModPlayer* player, ModPlayerPositionCallba
 /**
  * Set position (jump to specific pattern/row)
  */
-void mod_player_set_position(ModPlayer* player, uint8_t pattern, uint8_t row);
+void mod_player_set_position(ModPlayer* player, uint8_t pattern, uint16_t row);
 
 /**
  * Set BPM (125 is standard)
